% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cdc.samp.R
\docType{data}
\name{cdc.samp}
\alias{cdc.samp}
\title{A dataframe of 60 rows and 9 variables used to illustrate inference on
demographic characteristics}
\format{
A tibble with 60 rows and 9 variables:
\describe{
\item{\code{genhlth}}{Factor with levels \code{excellent}, \code{very good}
\code{good}, \code{fair}, \code{poor}}
\item{\code{exerany}}{Numeric vector; 1 if the respondent exercised in the
past month and 0 otherwise.}
\item{\code{hlthplan}}{Numeric vector; 1 if the respondent has some form
of health coverage and 0 otherwise.}
\item{\code{smoke100}}{Numeric; 1 if the respondent has smoked at least 100
cigarettes in their entire life and 0 otherwise.}
\item{\code{height}}{Numeric; respondent's height in inches.}
\item{\code{weight}}{Numeric;  respondent's weight in pounds.}
\item{\code{wtdesire}}{Numeric; respondent's desired weight in pounds.}
\item{\code{age}}{Numeric;  respondent's age in years.}
\item{\code{gender}}{Factor with two levels \code{m} \code{f}}
}
}
\source{
\url{http://www.openintro.org/stat/data/cdc.R}
}
\usage{
cdc.samp
}
\description{
This dataset is a random sample of 60 individuals from 20,000 responses in
the dataset cdc, itself a random sample from the year
2000 Behavioral Risk Factors Surveillance System (BRFSS)
(https://www.cdc.gov/brfss/index.html)
conducted by the US Centers for Disease Control. The dataset cdc
is available in this package and at OpenIntro.org. This random sample differs
from that posted on the OpenIntro labs,
https://www.openintro.org/data/index.php?data=cdc.samp.
}
\keyword{datasets}
