% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.mrna.R
\docType{data}
\name{coding.mrna.2024}
\alias{coding.mrna.2024}
\title{Protein coding transcript lengths for 99,775 genes retrieved 25 January 2024}
\format{
a dataframe with 79,105 rows and 5 columns:
\describe{
\item{\code{chromosome_name}}{a character vector, Ensemble label Chromosome/scaffold name}
\item{\code{external_gene_name}}{a character vector, Ensemble label Gene stable ID}
\item{\code{external_transcript_name}}{a character vector, Ensemble label
Transcript stable ID}
\item{\code{transcript_length}}{a numeric vector, Ensemble label Transcript length
(including UTRs and CDS)}
\item{\code{description}}{a character vector, Ensemble label Gene description}
}
}
\source{
{\url{http://www.ensembl.org/biomart/martview/}}
}
\usage{
coding.mrna.2024
}
\description{
Dataset with gene names, gene description, transcript name, transcript
length and chromosome location.  Data downloaded from Ensemble using Ensemble structured
Variables  have been renamed to match names in the 2018 data.
}
\keyword{datasets}
