% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-coding_mrna_2024.R
\docType{data}
\name{coding.mrna}
\alias{coding.mrna}
\alias{coding.mrna.2018}
\title{Protein coding transcript lengths for 79,105 genes, retrieved 27 January 2018}
\format{
a dataframe with 79,105 rows and 5 columns:
\describe{
\item{\code{chromosome_name}}{a character vector}
\item{\code{external_gene_name}}{a character vector}
\item{\code{external_transcript_name}}{a character vector}
\item{\code{transcript_length}}{a numeric vector}
\item{\code{description}}{a character vector}
}
}
\usage{
coding.mrna
}
\description{
Dataset with gene names, gene description, transcript name, transcript
length and chromosome location.  Data downloaded from Ensemble via a Bioconductor script
now obsolete.  Dataset has been retained in the package for compatibility with labs and
exercises in Introductory Statistics for the Life and Biomedical Sciences (ISLBS), Vu and
Harrington.  More current data is contained in coding.mrna.2024.
}
\keyword{datasets}
