% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/danish.ed.primary.R
\docType{data}
\name{danish.ed.primary}
\alias{danish.ed.primary}
\title{Data from the Danish study on triage in the emergency department (ED)}
\format{
A tibble with 6249 rows and 21 variables:
\describe{
\item{mort30}{numeric, 1 if patient died within 30 days of admission, 0
otherwise}
\item{triage}{factor, triage score given at arrival to ED.
Values green, yellow, orange, red, from lowest to highest priority
for treatment.  The value blue normally denotes severity not
warranting admission to the ED, but not participants coded blue
are in these data.}
\item{age}{numeric, age in years, rounded to lower integer}
\item{sex}{factor, female/male}
\item{albumin}{numeric, serum albumin, in g/L}
\item{creatinine}{numeric, serum creatinine, in umol/L}
\item{hemaglobin}{numeric, serum hemaglobin, in mmol/L }
\item{potassium}{numeric, serum potassium, in mmol/L}
\item{leuk.count}{blood leukocyte count, in 10E9/L}
\item{sodium}{numeric, serum sodium, in mmol/L}
\item{c.react.protein}{numeric, serum C-reactive protein}
\item{oxygen.sat}{numeric, peripheral arterial oxygen saturation, \%}
\item{resp.rate}{numeric, respiratory rate per minute}
\item{heart.rate}{numeric, heart rate, beats/min}
\item{systolic.bp}{numeric, systolic blood pressure, in mmHg}
\item{glasgow.coma.scale}{numeric, extent
of impaired consciousness in patients with acute medical condition or
trauma, scored between 3 and 15, 3 being the worst and 15 the best. Score
is based on 3 subscales, best eye, verbal and motor responses.}
\item{readmit.hosp}{factor, readmitted to hospital within 30 days,
yes/no}
\item{days.in.hosp}{numeric, number of days admitted to hospital}
\item{icu.time}{numeric, number of days in the intensive care unit.
value 99999 indicates patient not admitted to ICU}
\item{icu.status}{factor, patient admitted to ICU, yes/no}
}
}
\source{
\url{doi:10.5061/dryad.m2bq5}
}
\usage{
danish.ed.primary
}
\description{
Data from a prospective cohort study of triage scoring for an emergency
department (ED).  The study examined whether the use of patient level
measurements would improve an existing triage score. These data are the
training data (called primary data in the original manuscript) used for model building.
some variable names have been changed for readability, but the data on
21 variables for the 6,249 participants are otherwise unchanged.
}
\references{
{https://sjtrem.biomedcentral.com/articles/10.1186/s13049-017-0458-x?report=reader}
}
\keyword{datasets}
