% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-golub.R
\docType{data}
\name{golub.test}
\alias{golub.test}
\title{Ten cases randomly chosen from the larger \code{golub}
dataset to test a model for predicting leukemia type}
\format{
A tibble with 10 rows and 7135 variables
}
\source{
\url{doi:10.5061/dryad.m2bq5}
}
\usage{
golub.test
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{The first 6 variables contain data on clinical and
}\if{html}{\out{</div>}}

demographic features. The remaining 7129 contain normalized expression data
for the probes given by the variable name.  See the parent dataset documentation
for more information on the source and variable definitions.
}
\references{
{https://sjtrem.biomedcentral.com/articles/10.1186/s13049-017-0458-x?report=reader}
}
\keyword{datasets}
