% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golub.R
\docType{data}
\name{golub.train}
\alias{golub.train}
\title{A dataset of 62 cases randomly chosen from the larger \code{golub} dataset to build
a model for predicting leukemia type}
\format{
A tibble with 62 rows and 7135 variables
}
\source{
\url{doi:10.5061/dryad.m2bq5}
}
\usage{
golub.train
}
\description{
The larger golub dataset is a normalized version of the Golub leukemia data
from the \code{golubEsets} package,
available from \url{http://www.bioconductor.org/download/experiments/}.
Republished from the \code{hddplot} package by John Maindonald
on 10 March 2018.  The first 6 variables contain data on clinical and
demographic features. The remaining 7129 contain normalized expression data
for the probes given by the variable name.
}
\references{
{https://sjtrem.biomedcentral.com/articles/10.1186/s13049-017-0458-x?report=reader}
}
\keyword{datasets}
