% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperuricemia.samp.R
\docType{data}
\name{hyperuricemia.samp}
\alias{hyperuricemia.samp}
\title{Random sample of 500 cases from the hyperuricemia dataset.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 500 rows and 8 columns.
}
\source{
\url{doi:10.5061/dryad.n5j23}
}
\usage{
hyperuricemia.samp
}
\description{
Random sample from the hyperuricemia data.  See the larger
dataset for variable definitions.
}
\keyword{datasets}
