% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oibiostat-package.R
\docType{package}
\name{oibiostat-package}
\alias{oibiostat}
\alias{oibiostat-package}
\title{oibiostat: Data Package for OpenIntro Biostatistics, Introduction to Statistics for the Life and Biomedical Sciences}
\description{
This package contains the datasets used in the OpenIntro Biostatistics text, Introduction to Statistics for the Life and Biomedical Sciences.
}
\author{
\strong{Maintainer}: Dave Harrington \email{davidharrington@g.harvard.edu}

Authors:
\itemize{
  \item Morgan Breitmeyer \email{morganbreitmeyer@gmail.com}
  \item Julie Vu \email{julievu777@gmail.com}
}

}
\keyword{internal}
