% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-opp_insights_college.R
\docType{data}
\name{opp_insights_colleges_4year}
\alias{opp_insights_colleges_4year}
\title{Data from opp_insights_colleges that is restricted to 4-year, not-for-profit colleges}
\format{
A dataframe with 1285 rows and 26 variables
}
\usage{
opp_insights_colleges_4year
}
\description{
A dataset based on the larger Opportunity Insights data (opp_insights_colleges) that is
restricted to 4-year not-for-profit colleges.  Schools with 2-year programs or are
for-profit schools have been dropped from the parent dataset.  See the parent
dataset for source and variable documentation
}
\keyword{datasets}
