% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevend.samp.R
\docType{data}
\name{prevend.samp}
\alias{prevend.samp}
\title{Random sample of size 500 from the prevend dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 500 rows and 31 columns.
}
\usage{
prevend.samp
}
\description{
Random sample of size 500 from the 4,095 cases in the prevend dataset with all
31 variables.  See the parent dataset (prevend) for data source and variable
definitions
}
\references{
{https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0115755}
}
\keyword{datasets}
