% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.levels.R
\docType{data}
\name{sugar.levels.B}
\alias{sugar.levels.B}
\title{Simulated sugar (glucose) levels for 100 residents of a hypothetical neighborhood B}
\format{
A tibble with 100 rows and 1 variable
\describe{
}
}
\usage{
sugar.levels.B
}
\description{
Simulated sugar (glucose) levels for 100 residents of a hypothetical neighborhood B
}
\keyword{datasets}
